Minor refactoring and added some slash commands
